requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.8
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
pydantic==2.6.0
pydantic_settings==2.1.0
redis==5.0.1
hiredis==2.3.2
djangorestframework==3.14.0
transliterate-1.10.2